fix type errors in ABI

+++ precompiles/avs/events.go
@@ -271,7 +271,7 @@ func (p Precompile) EmitTaskSubmittedByOperator(ctx sdk.Context, stateDB vm.Stat
 	}
 	// Prepare the event data:sender,TaskResponse, BlsSignature, Phase
 	arguments := abi.Arguments{event.Inputs[2], event.Inputs[3], event.Inputs[4], event.Inputs[5]}
-	packed, err := arguments.Pack(params.CallerAddress.String(), params.TaskResponse, params.BlsSignature, params.Phase)
+	packed, err := arguments.Pack(params.CallerAddress.String(), params.TaskResponse, params.BlsSignature, uint8(params.Phase))
 	if err != nil {
 		return err
 	}
